captcha_model/requirements.txt
# PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Captcha Generation
captcha>=0.5.0

# Image Processing
numpy>=1.24.0
pillow>=10.0.0

# Configuration
pyyaml>=6.0

# ONNX Export and Inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Progress Bar
tqdm>=4.65.0
# 其他包

